requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
